split tests into 1:1 files
